+++ dbus-serialbattery/utils.py
@@ -421,9 +421,6 @@ SOC_LOW_ALARM: float = get_float_from_config("DEFAULT", "SOC_LOW_ALARM")
 BATTERY_CAPACITY: float = get_float_from_config("DEFAULT", "BATTERY_CAPACITY")
 INVERT_CURRENT_MEASUREMENT: int = get_int_from_config("DEFAULT", "INVERT_CURRENT_MEASUREMENT")
 
-# -- JK BMS settings
-JKBMS_CAN_CELL_COUNT: int = get_int_from_config("DEFAULT", "JKBMS_CAN_CELL_COUNT")
-
 # -- ESC GreenMeter and Lipro device settings
 GREENMETER_ADDRESS: int = get_int_from_config("DEFAULT", "GREENMETER_ADDRESS")
 LIPRO_START_ADDRESS: int = get_int_from_config("DEFAULT", "LIPRO_START_ADDRESS")